requirements.txt
gitpython
sentence-transformers
faiss-cpu
tqdm
python-magic
fastapi
uvicorn
requests